warnings: resolve compiler warnings on the release build
Get the x86-release build to zero warnings:

  - C4018 (signed/unsigned) in TilemapReader: explicit casts at the
    pattern_id / flags_count comparisons (semantics unchanged).
  - C4996 in GdiRender: fopen -> fopen_s in the debug BMP dump helper.
  - C4996 (strncpy) from the vendored Steamworks SDK headers: suppress
    via _CRT_SECURE_NO_WARNINGS on the four Steam TUs that include them
    (don't patch the SDK).
  - C4244 from vendored StapleGL headers: mark its include dir SYSTEM and
    add /external:W0 so third-party header noise stays out of our build.
  - C4244 wchar->char in ModLoader log lines (startup_scripts /
    execute_all_user_scripts_in): an explicit to_narrow() helper and a
    char-returning transform make the narrowing intentional.
  - FastRenderHooks is a WIP experiment (~30 warnings, incl. ones from
    template instantiations in STL/StapleGL headers that a per-TU /w
    can't reach): wrap the file in #pragma warning(push, 0) plus an
    explicit disable of 4244/4267/4018/4834 by number (the disabled-list
    survives the STL's own #pragma warning(push, 3)).

+++ src/ModLoader/TilemapReader.cpp
@@ -147,7 +147,7 @@ namespace rm_modloader::tilemap {
 	 */
 	void read_autotile(TilemapReader& reader, const glm::vec3& origin, const glm::vec2& tex_offset,
 		int pattern_id, std::span<const glm::vec2[4]> sources, int texture_index, const glm::vec2& animation_stride) {
-		if (pattern_id >= sources.size()) {
+		if (static_cast<size_t>(pattern_id) >= sources.size()) {
 			return;
 		}
 
@@ -468,7 +468,7 @@ namespace rm_modloader::tilemap {
 				for (size_t x = 0; x < width; ++x) {
 					size_t index = z * height * width + y * width + x;
 					int16_t tile_id = tilemap->map_data[index];
-					int16_t tile_flag = tile_id < tilemap->flags_count ? tilemap->flags[tile_id] : 0;
+					int16_t tile_flag = static_cast<DWORD>(tile_id) < tilemap->flags_count ? tilemap->flags[tile_id] : 0;
 
 					read_tile_primitives(reader, tile_id, tile_flag, glm::vec3(x * 32, y * 32, z_offset + z));
 				}
